requirements.txt
absl-py==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comtypes==1.4.11
contourpy==1.3.2
cryptography==45.0.7
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
eval_type_backport==0.3.1
exceptiongroup==1.3.1
fastapi==0.135.1
filelock==3.25.0
flatbuffers==25.12.19
fonttools==4.59.1
frozenlist==1.8.0
fsspec==2026.2.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
groq==1.0.0
grpcio==1.74.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.6.0
idna==3.11
importlib-metadata==6.11.0
invoke==2.2.1
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==2.12.0
kiwisolver==1.4.9
langchain==1.2.10
langchain-classic==1.0.2
langchain-community==0.4.1
langchain-core==1.2.17
langchain-mistralai==1.1.1
langchain-text-splitters==1.1.1
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.13
libclang==18.1.1
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.5
mdurl==0.1.2
mistralai==1.12.4
ml_dtypes==0.5.4
multidict==6.7.1
mypy_extensions==1.1.0
namex==0.1.0
narwhals==2.16.0
numpy==2.2.6
oauthlib==3.3.1
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
opt_einsum==3.4.0
optree==0.18.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==23.2
pandas==2.3.3
Pillow==9.5.0
propcache==0.4.1
protobuf==6.33.5
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
Pympler==1.1
pyparsing==3.2.3
pypiwin32==223
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pyttsx3==2.99
pytz==2025.2
pytz-deprecation-shim==0.1.0.post0
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.48
starlette==0.52.1
streamlit==1.54.0
tenacity==8.5.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.3.0
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==4.3.1
urllib3==2.6.3
uuid_utils==0.14.1
uv==0.10.8
validators==0.35.0
watchdog==6.0.0
Werkzeug==3.1.6
wrapt==1.14.2
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
